Drop usage of the `GleanTestRunner` from the tests
This uses, instead, the standard runner from Android tests.
