+++ samples/glean/build.gradle
@@ -32,7 +32,7 @@ android {
         versionCode 1
         versionName "1.0"
 
-        testInstrumentationRunner "org.mozilla.samples.glean.GleanTestRunner"
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
     }
 
     buildTypes {
